+++ tests/testthat/test-rate.R
@@ -46,9 +46,9 @@ test_that("Weighted rates produce known expected results", {
     expected = c(RR2w = 19.3 / 37.2)
   )
 
-  # '0' weights
+  # individual zero weights are valid (for example, in two-phase designs)
   w2[5] <- 0
-  expect_warning(outcomerate(x1, weight = w2), regexp = "zero")
+  expect_warning(outcomerate(x1, weight = w2), NA)
 })
 
 test_that("Should return numerator and denominator if asked", {